seabios: Remove PREMIRRORS statement.
First, PREMIRRORS is a little too strong to be defined at the recipe
level like so. I imagine it is more intended at providing a local.conf
override to fetch sources from a local cache. Having PREMIRRORS_prepend
makes that much more difficult.

Also, in this case, the upstream recipe has been fixed (at least was
yesterday).

This source is quite volatile. It should be recommended to fetch from a
stable mirror.

Maybe put the tarball on mirror.openxt.org and add a {PRE,}MIRRORS
statement in the default configuration file.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1410