module-signing: Fix SIGN_FILE error detection
When SIGN_FILE cannot be found, we should fail the signing function.
The existing check is wrong since SIGN_FILE will be set to the last
value of the for loop, even if it is not executable.  Repeat the
execution check a second time to see if we should bail.

This was introduced as originally there was a second variable that
assigned to SIGN_FILE.

OXT-1403

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>