xen: Amend xen.efi LoadOptions handling.
A more thorough sanity check is required in get_argv() to determine the
content of LoadOptions and its relevance to the cmdline:

Here is an example of LoadOptions returned by EFI boot-services on two
different platforms:

On a Dell 7450:
>00000000  01 00 00 00 62 00 4f 00  70 00 65 00 6e 00 58 00  |....b.O.p.e.n.X.|
>00000010  54 00 00 00 04 01 2a 00  01 00 00 00 00 08 00 00  |T.....*.........|
This content would have get_argv() return argc = 1, while nothing is
really usable to the cmdline.
In turn, efi_arch_handle_cmdline() would not prepend "xen" to the
cmdline set in the MBI structure.

On the other hand, on an HP 800 G1:
>00000000  d0 9b 00 f0 d0 9b 00 f0  c3 e2 00 f0 d0 9b 00 f0  |................|
>00000010  d0 9b 00 f0 54 ff 00 f0  47 66 00 f0 d0 9b 00 f0  |....T...Gf......|
This content would have get_argv() return argc = 0, which is legitimate.
In turn, efi_arch_handle_cmdline() would prepend "xen" to the cmdline
set in the MBI structure then passed to TBoot, in OpenXT.

This behavior will result in TBoot measurement failing on PCR18 when
performing forward sealing as the command-line used by the forward
sealing scripts will not include the prepended "xen " statement.

Per the author of the patch, efi-load-option-support.patch was an early
iteration of a patch now upstreamed, which can be backported in its
place and handles the LoadOptions with more consistency.

Now efi_arch_handle_cmdline() will, in both cases, prepend "xen" to the
MBI cmdline. Since this is not the desired behavior on OpenXT EFI/DRTM
bootstrap, the image name is not set like that, a second patch will only
prepend the image name to the command-line if it is set in LoadOptions,
and not rely on a default value.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1432