linux: Remove break-8021d.patch
break-8021d only operates on 01-80-C2-00-00-00 Spanning Tree
(STP/RSPT/MSTP), so it was broken for forwarding 802.1X PAE.  Since
break-8021d was written, upstream linux has gained similiar
functionality in the sysfs group_fwd_mask.  That allows forwarding most
of the same traffic, but a few packet types are not allowed through:
01-80-C2-00-00-00 Bridge Group Address used for STP
01-80-C2-00-00-01 (MAC Control) 802.3 used for MAC PAUSE
01-80-C2-00-00-02 (Link Aggregation) 802.3ad

01-80-C2-00-00-00 can also be allowed by disabling STP.

The only user of break8021d was the linux stubdom, but that was switched
to using group_fwd_mask.  Network slave in NDVM disables stp on bridges it
creates and does not use break_8021d.

OXT-1437

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>