xenclient-idl: Add recipe version
xenclient-idl lacked a version since it was a git recipe, so bitbake
assigned it version git-r0.  When idl.git changed, dependency tracking
was thrown off since the version remained the same.  This would break
rebuilds of dependant packages since their sysroots would not be updated
with the newer version.

Set PV so the recipe is versioned and dependencies can be tracked.

OXT-1436

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>