uefi/tboot: Load microcode from MBI2 modules.
During UEFI/TBoot boostrap, xen.efi will fetch the microcode using the
openxt.cfg configuration file and write the binary blob in the last MBI
module[1], then passes everything to TBoot.

Once TBoot is done, the new measured instance of Xen starts and boots
from the MBI2 information. The remaining modules include the microcode,
but the current Xen cmdline does not point which one is the
microcode.

Add ucode=-1, the microcode being guaranteed to be the last module.

[1] https://xenbits.xen.org/gitweb/?p=xen.git;a=blob;f=xen/arch/x86/efi/efi-boot.h;h=5789d2cb70726a556785e823ee4eafcdb514aa6c;hb=HEAD#l287

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1452