openxt_image_types: Fix .vhd size calculation
The VHD size calculation was mistakenly using the wrong units.  This
meant that 2GB VHDs were being generated for our images.

ROOTFS_SIZE is in KB, so ALIGN should be 2 * 1024 for 2MB in KB.  The
TARGET_VHD_SIZE_KB was mislabeled since it is really MB which is the
unit that `vhd convert` expects.  Rename the two TARGET_VHD_SIZE*
variables so the units are clear.

Finally, correct the alignment calculation.  We want to add $ALIGN - 1
to force ROOTFS_SIZE up to the next chunk for any partially chunks.
Adding $ALIGN + 1 always forces up one chunk and potentially two.

A .disk partitioned disk image will be larger that ROOTFS_SIZE
kilobytes.  Now that we are using a more-precise .vhd size, we need to
take into account the extra size of the .disk.

We check the read size against ROOTFS_SIZE and adjust accordingly.
Maybe we should just always use the READ_SIZE_KB?

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>