[recipes-openxt] remove script guest-process-stats
This script appears to be part of a partial package that was not fully exported
in the open sourcing of XenClientXT. This scripted is called by another script
but there is no recipe that installs. In addition, this script is in fact a
Ruby script which is not installed or part of OpenXT. As a result this script
is just being removed.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>