tpm2-tss: Let OE handle the autoconf.
tpm2-tss ships with a "bootstrap" script to be run to configure the
build environment.
The script does two things:
- Create an src_vars.mk file that will generate the values of automake
variables for the source files being built
- Install, with symlinks, missing autotool files (autoreconf).

While the first one is harmless, the second is not intended in a
Bitbake/OE environment, as autotools.bbclass will take care of doing
that.

Patch out the autoconf from the bootstrap script.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1485