rsyslog: Drop klog messages from VMs
We get duplicate output in /var/log/messages since the linux kernel
prints to /dev/hvc0 and then rsyslog prints the message again with the
kernel: prefix.  Syslog in dom0 catches all the messages since it is
running for the entire lifetime of our VMs.  Remove klog from
rsyslog.conf to drop the duplicate messages.

Dom0 and the installer keep the line since it is needed for their kernel
messages to get logged.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>