rsyslog: Split conf files into seperate subpackage
Split rsyslog.conf into rsyslog-conf and make rsyslog RRECOMMEND that
sub-package.  This will result in rsyslog-conf being installed by
default, but not installed when another package (rsyslog-conf-dom0)
RPROVIDEs rsyslog-conf and is explicitly installed.

RDEPENDS rsyslog-conf results in a installation conflict when also
installing rsyslog-conf-dom0, so that approach is not viable.

At the same time, we update the packagegroups for dom0 and installer to
ensure the desired configs are installed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>