tpm2: Store pcr configuration in ${tss}.pcrs
Instead of hardcoding the pcr configuration into scripts, store the
configuration into the ${tss}.pcrs file.  This way we remove hardcoding
in our scripts.  Also, it will enable the tpm-quirks functionality to
tweak the pcr config.

TPM 1.2 sealing structure incorporates the PCR configuration.  Not sure
why TPM 2.0 does not.  Anyway, it's not sensitive, so it's okay to
store.  If an attacker changes it, they still cannot unseal.

OXT-1594

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>