tpm-scripts: Re-enable pcr1 quirk
We have a Dell Laptop where the USB-3 Dock changes PCR1 depending on
whether or not it is plugged in.  So by including PCR1 in the sealing
set, the system can only be boot in one of docked or undocked.  Dropping
PCR1 from the sealing set lets it boot either way.

The Dock changes 2 items in the PCR1 event log.  The first is the
measurement of the BootOrder UEFI variable (event 0x80000002) changing
with a the addition of a new BootVariable.  The second is the new boot
variable for "USB NIC" added into to the event log, again event
0x80000002.

Drop the trailing space off the sed line.  It's a corner case, but it
would not match if '-p 1' were the end of the line.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>