rsyslog: Fix broken packaging
I had $(sysconfdir} - note the erroneous parenthesis - so it wasn't
expanding the CONFFILES properly.  FILES was also not getting set since
${CONFFILES}_${PN} wasn't getting expanded.  This meant
/etc/rsyslog.conf was still packaged in "rsyslog".

However, /etc/rsyslog.conf was still getting installed from
rsyslog-conf-dom0.  /etc/rsyslog.conf was marked as a CONFFILE even in
the "rsyslog" package.  So when you installed rsyslog-conf-dom0, it
would rename the original to /etc/rsyslog.conf-opkg and then install the
new conf file as /etc/rsyslog.conf.  Dom0 and installer have those extra
-opkg files hanging around.

This fixes up the recipe to set FILES and CONFFILES properly, but it
doesn't address the -opkg situation.  I'm not sure how to tackle that.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>