rsyslog-conf-dom0: Fix RPROVIDES
Fix the typo-ed ${PN} so RPROVIDES is set properly.  This prevents the
duplicate /etc/rsyslog.conf installation since rsyslog-conf is no longer
installed.

Also add RCONFLICTS in an attempt to avoid duplicate /etc/rsyslog.conf
installation.  Without it, I haven't seen the -opkg overwrite, so it may
not be strictly necessary.  But it seems like the appropriate entry
since we only want on install.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>