module-signing: Re-work sign-file selection
On a Linux uprev, module signing can fail on a dirty build when
sign-file is called with a rpath pointing into a modules sysroot.  The
module's sysroot path will be gone, so sign-file fails to run.
/home/build/openxt-master/build/tmp-glibc/work-shared/xenclient-dom0/kernel-build-artifacts/scripts/sign-file: error while loading shared libraries: libcrypto.so.1.0.2: cannot open shared object file: No such file or directory

For a kernel build, we should run ${B}/scripts/sign-file.  Instead of
searching for this at runtime, use a variable to set the the SIGN_FILE
location for modules and override for a kernel build.

While doing this, move 'require classes/module-signing.bbclass' into
kernel-signing.bbclass and rename to kernel-module-signing.bbclass.
That consolidates the functionality required for the kernel in one
place.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>