mod-sign: Add dep on later task.
This is mostly noticeable if KERNEL_MODULE_SIG_KEY is not provided.
Refreshing a patch, or changing kernel revision, will not always have
the out-of-tree modules be signed with the new key.

Add a dependency on the kernel do_install task to have the new key files
present in the shared workdir when running sign_modules and re-run
do_sign_modules if virtual/kernel:do_install was done since last
signing.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>