Add tag MULTIBOOT2_TAG_TYPE_EFI64 from Xen EFI to TBOOT
The multiboot header going from Xen EFI to TBOOT was missing the the MULTIBOOT2_TAG_TYPE_EFI64 tag, which is a pointer to the EFI System Table. TBOOT would fail to load the ACPI tables from EFI thereby causing a failure when going to SENTER. With this tag TBOOT now recognizes the preload as EFI and loads the proper tables including the Root System Description Table (RSDT). I have tested this on both HP and Dell computers

Signed-off-by: Richard Turner <turnerr@ainfosec.com>