Only set rtc for dom0 & installer MACHINE_FEATURES
dom0 and installer are the only machines that actually talk to a
hardware real-time clock.  The other virtual machines don't need to work
with an rtc, so it can be remove.  To actually do so, we need to add it
to MACHINE_FEATURES_BACKFILL_CONSIDERED since it defaults enabled when
unset.  With that we need to explicitly add it for dom0 and installer.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
(cherry picked from commit 662797fbffc8a697289a6d481f1014efa36451d6)