cross recipe: renaming initramfs.gz to initrd
OpenEmbedded classes and recipes follow the convention that an image's
init image is called initrd when built into the image. In order to
better use OpenEmbedded's existing tooling it is helpful to follow then
naming convention of the init image.

This is a cross cutting change that must be made in a single commit. As
a result this make the appropriate changes across the supporting files
for both openxt-measuredlaunch and xenclient-dom0-tweaks.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>