classes: expand the openxt disk type
This refactoring and expansion was done to mimic the upstream ability to
generate both syslinux and efi versions of a hard disk image. The upstream
versions has portions of the processing written in python which is not
compatible with the chainable CONVERSION process. Alignment to the CONVERSION
proces is necessary since OpenXT image types are written as CONVERSION
functions.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>