class: fixing a remnant v4v reference
The class refactoring patch set started prior to the switch over to
Argo. As a result the `read_only_rootfs_hook_extend` rootfs postprocess
command was still looking to rewrite the sshd configuration for v4v
instead of argo. The result is that `sshd-argo` used the standard
`sshd-config` which configured it to listen on argo port 22 instead of
port 2222.

This fixes by replacing the v4v portion of names to argo.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>