ml: PCR IDs conf files without options.
Currently, PCR IDs used for sealing are saved in:
/boot/system/tpm/config.tss.pcrs
/config/config.pcrs

Both files are used by the sealing/unsealing processes and had the files
contain the options to be passed to the TSS utilities.

Change that behavior so the files are now list of PCR IDs separated by
spaces. The scripts are responsible to assemble the correct options when
invoking the TSS utilities or other script helpers.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1664