stubdoms: introducing stubdom recipes
Stubdomain images must be include into the domain that is resposible
for starting the domains. The issue is that stubdomains are built using
a different MACHINE type than the images that will contain the image.
This means they cannot be included in the same manner as initramfs
images. In these recipes an approach is introduced to provide a method
to copy the studomain image and kernel into the image at build time.
The recipe approach is being used as it provides for the target image
recipe to detect when the stubdomain image has been rebuilt.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>