xen: Remove hotplug-status writes from libxl
vfb doesn't care about hotplug-status, so it can be dropped.  The vif
script handles it for vif/vwif, so that can be dropped.

/etc/xen/script/block handles writing it for block devces, but the iso
hotswap path does a lot of xenstore re-writing.  tapback needs it set to
function properly.  Testing shows the libxl write to hotplug-status is
unnecessary.  The block script is called, so hotplug-status is written
that way.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>