qemu-dm: chardev-argo: Use DOMID_FIRST_RESERVED
The domid comes in as a 64bit integer, so we want to sanity check the
remote domain we will connect to.  Invalid domids are >=
DOMID_FIRST_RESERVED (0x7ff0).  Use the define provided in xen/xen.h,
which we already include, to ensure that the domid option is an
allowable domid.  Raw values 0x7ff4(DOMID_INVALID) or the earlier
0x7fff(DOMID_IDLE) were both incorrect since some of the reserved domids
were still allowed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>