initramfs-framework: adding functions module
This introduces the function framework module that provides utility
functions that can be used by other modules. The first of two functions
introduced here is the capability to break out to a shell before and
after a module is ran. The second function is a log function that
enables sending severity labeled messages to the kernel message log.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>