xen-libxl: Use populate-volatiles script.
Upstream xl will use xencommons to create /var/{run,log}/xen
directories at boot. With SELinux, this requires labeling the files in
the non-selinux aware scripts.

Instead, use populate-volatiles.sh to do so as it will call restorecon
on the created objects automatically.

The scripts were originally created in mountall.sh which is run at a
lower priority than populate-volatiles.sh, so it is not a problem to
move the creation earlier.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1682