seabios: Handle pre-deployed ROMs for now.
Seabios dropped support for the compile time option:
9d691ac optionroms: Drop support for CONFIG_OPTIONROMS_DEPLOYED

This is no longer necessary as seabios will handle copying the VBIOS
exposed by the main VGA PCI device to 0xc0000 for us. Now it can be
immediately available from a PCI device (PCI PT in OpenXT's case), or
because QEMU loaded the ROM for the emulation. In both case, SeaBIOS can
handled it.

Since the ROM file has to be available to QEMU in the case of
stdvga/cirrus emulation, it does mean it will be both in dom0 and the
stubdomain initramfs to support stubdomain/non-stubdomain HVM guests.
The VGABIOS ROM built by OpenXT being <50KB, this is not a big deal.

This also makes it easier to switch the ROM file for one compiled with
DEBUG.

This patch was also submitted upstream years ago:
https://seabios.seabios.narkive.com/lEgdFhdv/patch-0-3-use-vga-hooks-to-make-xen-vga-passthrough-work-on-certain-devices

Until QEMU is configured to emulate the ROMs from files so SeaBIOS can
do the copy to memory, shortcut the PCI ROM handling and use what
hvmloader pre-deployed.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>