xen: Work-around vbd3 support with udev.
Blktap3 tapback expects "hotplug-status" to be switched to "connected"
by the hotplug scripts before it initiates the connection with the
frontend. This used to be handled by a patch against libxl prior:
    31f6540c xen: Remove hotplug-status writes from libxl
This action was moved to the block script provided by the toolstack
(https://github.com/OpenXT/toolstack/pull/37), which this changeset
intend to no longer use.

The upstream hotplug scripts do not handle vbd3 without minor changes
and existing support in libxl would require even more modifications.
Until this is all simplified and refactored, use a custom hotplug
script that is called by a udev rule to set the hotplug-status using the
upstream hotplug script framework.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1356