xen: Add missing dummy XSM code to init quiet Argo XSM check
The recent change to quiet the AVC issued when a domain does not have
XSM argo enabled unfortunately omitted a change to add initialization
logic to the dummy policy, so when running with the dummy policy active,
it results in a crash in Xen on boot.
This change adds the needed initialization.

Fixes OXT-1692.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>