refpolicy-mcs: Derive instead of bbappend.
Different images may require different modules included in the final
policy. In order to achieve this, similar policies will have to be built
and packaged separately using different recipes. This is until some
facility allows to handle semodule use in read-only image creation.

This change abandons the .bbappend approach for the refpolicy-mcs and in
turn introduces a policy derived from meta-selinux refpolicy (includes
as much as possible from it), but has its own recipe.

Re-use the .inc files from meta-selinux to share as much configuration
as possible. This requires customizable_types, setrans-mcs.conf,
setrans-mls.conf to be copied to the layer.

The versions of the policies follow the one in meta-selinux.  Patches
and policy-files are moved to a version specific directory in the layer,
similar to how it was before, without the globbing.

Classes and existing recipes are changed to include the name of the new
policy.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1695