xc-toolstack: Skip file-rdeps QA check.
Bitbake/OE is able to detect that nothing RDEPEND on the runtime for
intermediate .cma/.cmi files shipped in -dev packages.
Since OpenXT does not plan to provide an SDK and will use -dev with the
cross-toolchain, just disable the QA check for now.

Note: That QA test did not trigger back on pyro.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>