xen-ocaml: Use native cross-compiler.
By default, the Xen build-system will use ocamlopt which is the bytecode
ran with the ocamlrun from the cross toolchain. While the
cross-toolchain should not provide an ocamlrun for a start, even when
using the host ocamlrun (ocaml-native), ocamlopt seems to fail to link
with the libraries present in the sysroot (libm.so.6, libmvec.so.1,
libmvec_nonshared.a).

In any case, the native ocamlopt.opt seems fine, so use that.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>