trousers: recipe overhaul.
Apply https://www.openembedded.org/wiki/Styleguide.

Assign the tss:tss correct ownership on files for tcsd. The daemon is
ran under tss:tss and needs access to these.

Restore trousers PS-file override:

system.data is supposed to be created when taking ownership of the TPM.
It then serves as a system scope persistent storage for keys that are
not stored on the TPM.

TCSD will not start if the file does not exist and it will try to change
it when not to its liking, which is a problem for installer->dom0(RO) or
even initramfs->dom0(RO).

OpenXT does not use that storage though. Historically the file is stored
in the boot partition, so tcsd finds system.data.auth on first boot,
although the installer took ownership of the TPM...  The standalone
tools somewhat mitigate that, so initramfs gets away with it (even if
the PS file isn't available weirdly).

This is spread across the platform: tpm_ownership in the installer, tpm
operations in the initramfs, then in the pre-init init.root-ro, then
again once the bootstrap completes and tcsd starts.

It should be overhauled, but TPM 1.2 is legacy.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>