xen: Remove duplicate xenstore writes
libxl_create.c:libxl__domain_make already creates these xenstore entries
down on lines 725-736.  The one change is that we revert to the default
read-only permissions on control.  This is fine since the sub-nodes
inside control are created read-write for the domain.

The "error" and "messages" nodes are dropped by this patch.  They are
unused as far as I can tell - a relic of xenvm which created them.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>