uim: Port to upstream recipe.
Remove portable-object files from the layer, these should not be
necessary.
Also remove the LINGUAS leftover from XenClient, this should be handled
downstream mostly.

Agent:
Handle gtk3 version of the toolbar. Although it is not the one used in
UIVM, it is conceivable to upgrade to it at some point.

Patch-queue:
toolbar_new() now calls button_create() which already has the branding
for OpenXT.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>