lvm2: Configuration depending on image.
OpenXT's initramfs does not provide udevd, so LVM cannot obtain the
device list through it, which will have to timeout to complete boot.
Backport the patch to avoid the timeout on libudev calls as well.

Since lvm2 was upgraded, refresh the configuration file now that it is
modified from the example.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>