surf: Remove unused and handle RO filesystem.
Surf 2.0 recipe was appended to provide a .desktop file which is not
loaded in the current configuration. Since the recipe is maintained in
this layer and not backported from another, remove the bbappend and the
.desktop file.

Surf needs a volatile directory for its own use. Provide the necessary
configuration to have /root/.surf, since UIVM runs in Kiosk mode.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>