resized: Replace with an agent script.
resized is a daemon in UIVM watching the XenStore key:
/xc_tools/switcher/current_display_size
and using libxrandr to create a new mode to be added to the dummy output
added in xf86-video-fbdev Xorg driver.

Resized is ~650 lines of C, yet UIVM still installs the xrandr utility
which can already do all that.
resize-agent is a replacement for resized using xrandr and is around
100 lines of bash script.

Agents are currently packaged with xenclient-uivm-xsessionconfig, which
is not great. Fix the RDEPENDS nevertheless until that is changed.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1724