machines: Consolidate DEFAULTTUNE
Set these common values in xenclient-common.conf and remove the
duplication.  Thie make the commonality of all the machines explicit.

This changes xenclient-syncui to core2-64 from the unspecified default
of core2-32.  It's unclear if xenclient-syncui is even used anymore.
do_build.sh shows it used to build sync-wui which is a Web UI, so the
build architecture shouldn't matter.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>