xf86-fbdev: Resize shadowFB on CRTC resize.
OpenXT adds an output/CRTC to support XRandR modeline to be added and
used for display. When the resize event occurs, the shadowFB is not
resized. This mostly goes unoticed as 0.4 and previous versions of the
XOrg driver were over-allocating the shadowFB, leaving plenty of space
even for very large modes.

Take the opportunity to cleanup the patch, add TRACE macro to track down
problem, should they arise, and use a consistent code-style in the
patch.

Remove unused chunks of the patch regarding modeline validation, this
has no effect.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1723