xen: Patch block-tap script
This brings in a patched block-tap script.  It provides the device
sharing checks previously done in libxl.

Calling tap-ctl passes an open file descriptor to tapback which messes
up locking.  That is why we have `eval "exec $_lockfd>&-` to close the
flock fd.

It relies on TAPDISK3_CRYPTO_KEYDIR being set in the environment by
xenmgr to allow tapdisk to unlock encrypted VHDs.

Remove modprobe of blktap.  It can only be built-in, so it is pointless
to modprobe.  This also avoids xend_t needing modprobe domtrans for the
SELinux permissions.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>