Upgrade dojo from 1.7.2 to 1.7.12
1.7.12 is the latest micro release of the 1.7 line.
It was released in March 2017 (vs May 2013 for 1.7.2).
One notable improvement is that 1.7.12 fixes scroll-down lists
with more recent libwebkitgtk versions.

1.8 releases break the UI, so minor upgrades are not an option.
More specifically (thanks @jandryuk):
The broken UI fails to render icons along the top bar.
Instead it prints the text labels oriented vertically.
The labels cannot be clicked.
The lower status bar doesn't render the measured launch or power icons.
A VM can be created through the wizard,
but the details button cannot be clicked on the resulting VM.

Signed-off-by: Jed <lejosnej@ainfosec.com>