linux: i915 ioctls have to be contiguous.
i915_ioctls is an array of struct drm_ioctl_desc, which size is
stored in struct drm_driver driver and calculated using the ARRAY_SIZE
macro. Each ioctl is initialized in the array using DRM_IOCTL_DEF_DRV.
In the case ioctls are not contiguous, it will create empty entries in
the array that can be reached in drm_ioctl.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1730