qemu-dm: acpi-pm: Use device-model xenstore directory
6dde8833f7de "xen: Remove duplicate xenstore writes" changed a domain's
control xenstore directory to read-only.  This broke stubdom qemu's
ability to write to "control/hvm-powerbutton-enable".  Without that
node, xenmgr wouldn't use the acpi shutdown paths.

Switch to writing hvm-powerbutton-enable and hvm-shutdown in the
device-model directory of qemu-dm's domain.  That is either the stubdom
or dom0.  Logically it makes sense to put this device-model information
underneath the device-model instead of the guest.  It also side steps
the permission issue.

This needs a corresponding change to xenmgr.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>