Networkmanager: Fix issues related to devices managed
NetworkManager.conf: Configuration filters the devices to be managed by networkmanager.
		     We don't want all the devices to be managed (for ex: brshared, brany, etc.., ).

auto-dhcp-for-eth0.patch: Since eth0 is added as a bridge slave for brbridged, dhcp is disabled by default on eth0, this patch enables it.

fix-device-type-property-exposure.patch: This fixes the device type property exposed to clients like nm-applet, nmcli, etc..,

OXT-1281

Signed-off-by: Mahantesh Salimath <salimathm@ainfosec.com>