xenclient-installer-part2-image: satisfy prelink need for image /etc
The `pre_image_mangle` task removes some unpopulated directories from
the rootfs directory. Unfortunately the standard image prelink task
assumes the presence of the etc directory in the image rootfs and fails
when it is absent.

To satisfy this requirement and unblock the build, bracket the
prelink_image task with prepend/append to temporarily supply a
${IMAGE_ROOTFS}/etc directory.

OXT-1751

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>