grub-efi: Amend FILESEXTRAPATH to use local cfg
The cfg file is used by grub-mkimage to let grub know where to find its
configuration. For the iso installer, the default one pointing to the
ESP for grub.cfg will not work.

Without the correct cfg file passed in SRC_URI, booting from a USB stick
that was created from the installer ISO will boot to a grub prompt.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>