tpm-functions: Fix tpm_is_owned for tpm2
This function, on an un-owned tpm2, throws an error:
[: -eq: unary operator expected

When the tpm2 is un-owned, `grep -q` return 1.  The conditional is
skipped, and state is still "" when it gets to the inversion on the way
out of the function.  Initialize the state to 0 to avoid this issue.

The tpm1 path is fine since it always sets state.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>