initscripts: s/tty0/console in populate-volatiles.sh
/dev/tty0 is only present when a kernel is configured with
CONFIG_VT_CONSOLE, so the output redirection is incorrect when /dev/tty0
is not present.  A regular file will be created as /dev/tty0.

Use /dev/tty since "it is a synonym for the controlling
terminal of a process, if any." (from tty(4) manual page).  But that may
not be present either.  Use /dev/console which as the system console
makes sense for printing these messages.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>