dbd: Unify and rewrite recipe
dbd and db-tools are both built from the same repository and from the
same path in the directory. The OCaml toolchain does not separate source
and build artifacts. This leads to a world of troubles when using
externalsrc since both rely on the local findlib database.

There is little reason dbd and db-tools are built by separate recipes.
Unify all that and take the opportunity to fix the following:
- db-cat is overwritten by db-write due to a typo in the recipe
- remove haskell.bbclass, nothing uses it
- Use B where appropriate, but...
- brokensep (B = ${S}) is still required as the Makefile references
  autogen directly
- Fix RDEPENDS on libargo

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>