xenclient-toolstack: Overhaul the recipe
The xenclient-toolstack recipe is quite confusing as it stands. The only
built components it provides are ocaml libraries. The rest are scripts,
most of which are either not installed or not deployed in the end.

The MACHINE specificity seems to only stand on an RDEPENDS requirement
which is no ground for that sort of configuration, this can be handled
directly by each image configuration, and an image specific script.

Introduce openxt-ocaml-libs that will build the ocaml libraries that
other projects depend upon. Since ocaml statically link, this should
only end up in DEPENDS.

Introduce xen-<type>-scripts recipes that will move the scripts to
the layer. xen-vif-scripts-{dom0,ndvm} takes care of the image
specificity. the block and tap scripts are not used by blktap3, so not
deployed in common installations, but are provided for people interested
in blktap2.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>