distro openxt-main.conf : enable multilib with BASE_LIB definitions
Enable multilib to support building Xen's hvmloader with a 32-bit
sysroot for materials needed by the build.

Set the lib directory for 64-bit to /lib and to /lib32 for 32-bit.

When multilib is enabled, the default for 64-bit is /lib64 and
32-bit is /lib, but the current OpenXT haskell compiler breaks
in that configuration: the compiler package ghc-native builds to
completion but all compilation of haskell packages using it fails.

This is avoided by putting the 64-bit libraries into /lib instead,
so do that.

OXT-1694

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>