installer: use do_deploy instead of postprocess
Recipes for images should not have an SRC_URI or process local files.
This is supposed to be handled by packages and the sstate. Instead
deploy.bbclass provides an abstract way to stage files to be deployed on
bootable media.

Append syslinux, grub-efi and xenclient-installer, acms, tboot to have
each implement, or extend the existing, do_deploy task and have it ship
the files necessary to assemble the bootable media.

This replaces the use of IMAGE_POSTPROCESS in xenclient-installer-image
and leverages the sstate for the deployed resources.

Existing trees merging this changeset might have to clean the following:
- deploy/images/openxt-installer/iso/isolinux.cfg,
- deploy/images/openxt-installer/iso/bootmsg.txt,
- deploy/images/openxt-installer/netboot/pxelinux.cfg,
- deploy/images/openxt-installer/tboot.gz
- deploy/images/openxt-installer/*.acm
- deploy/images/openxt-installer/microcode*.bin
- deploy/images/openxt-installer/grub*.efi

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>