linux-openxt: blktap2: convert to blk-mq for 5.4
Initial blk-mq conversion.  This is modeled after xen-blkfront and
simpler drivers like sunvdc.

tap->tag_set.queue_depth is set to BLKTAP_RING_SIZE / 2 which matches
xen-blkfront.  The /2 is to handle 64k pages, which we don't have to
worry about, so it's probably unnecessary.  But it's also safer, so it
is set for now.

There are some minor API fixups for 5.4 in addition to the blk-mq
conversion.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>